create abstract average class
